update: asset changes and bug fixes

+++ Assets/Scripts/LectureReference.cs
@@ -8,16 +8,18 @@ public class LectureReference : MonoBehaviour
     public Lecture lectureInfo;
     private StageManager stageManager;
 
-    public void Start() {
+    public void Start() 
+    {
         stageManager = GameObject.FindGameObjectWithTag("MainScript").GetComponent<StageManager>();
         gameObject.GetComponent<Button>().onClick.AddListener(OnButtonPress);
     }
 
-    public void OnButtonPress() {
+    public void OnButtonPress() 
+    {
         stageManager.levelSelectUi.SetActive(false);
         stageManager.lectureUi.SetActive(true);
 
-        GameObject currentLecture = stageManager.lectureUi.transform.GetChild(lectureInfo.lectureType).gameObject;
+        GameObject currentLecture = stageManager.lectureUi.transform.GetChild((int)lectureInfo.lectureType).gameObject;
         currentLecture.SetActive(true);
 
         LectureManager lectureManager = currentLecture.GetComponent<LectureManager>();
